feature: make pages constructor nonFound404
