+++ src/components/app/app.module.css
@@ -8,11 +8,3 @@
   align-items: center;
 }
 
-.main{
-  height: calc(100% - 88px);
-  width: 1240px;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
