+++ src/pages/constructor.module.css
@@ -1,0 +1,8 @@
+.main{
+    height: calc(100% - 88px);
+    width: 1240px;
+    margin: 0 auto;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+  }
