+++ src/pages/notFound404.module.css
@@ -1,0 +1,30 @@
+.main{
+  height: calc(100vh - 88px);
+  width: 1240px;
+  margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.container{
+  max-width: 720px;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.title{
+  text-align: center;
+}
+
+.link{
+  text-decoration: none;
+  color: var(--text-primary-color);
+  border-bottom: 2px solid rgba(255, 255, 255, 0);
+  transition: .3s;
+}
+.link:hover{
+  border-bottom-color: rgba(255, 255, 255, 1);
+}
